# Conflicts:
#	app/src/main/java/com/worldventures/dreamtrips/core/repository/SnappyRepositoryImpl.java
#	app/src/main/java/com/worldventures/dreamtrips/wallet/di/WalletActivityModule.java
#	app/src/main/java/com/worldventures/dreamtrips/wallet/di/WalletCommandModule.java
#	app/src/main/java/com/worldventures/dreamtrips/wallet/service/command/SyncCardsCommand.java
#	app/src/main/java/com/worldventures/dreamtrips/wallet/service/command/reset/RemoveSmartCardDataCommand.java

+++ app/src/main/java/com/worldventures/dreamtrips/wallet/service/command/reset/RemoveSmartCardDataCommand.java
@@ -23,6 +23,12 @@ public class RemoveSmartCardDataCommand extends Command<Void> implements Injecta
    @Inject LostCardRepository lostCardRepository;
    @Inject SmartCardInteractor smartCardInteractor;
 
+   private final boolean withPaymentCards;
+
+   public RemoveSmartCardDataCommand(boolean withPaymentCards) {
+      this.withPaymentCards = withPaymentCards;
+   }
+
    @Override
    protected void run(CommandCallback<Void> callback) throws Throwable {
       smartCardInteractor.activeSmartCardPipe()
@@ -34,9 +40,11 @@ public class RemoveSmartCardDataCommand extends Command<Void> implements Injecta
    private Observable<Void> removeCache() {
       return Observable.defer(() -> {
          try {
-            persistentWalletCardsStorage.deleteWalletCardList();
-            persistentWalletCardsStorage.deleteWalletDefaultCardId();
-            cardListStorage.deleteWalletCardList();
+            if(withPaymentCards) {
+               cardListStorage.deleteWalletCardList();
+               persistentWalletCardsStorage.deleteWalletCardList();
+               persistentWalletCardsStorage.deleteWalletDefaultCardId();
+            }
             snappyRepository.deleteSmartCardFirmware();
             snappyRepository.deleteSmartCardDetails();
             snappyRepository.deleteSmartCard();
